dagster_university/dagster_and_etl/pyproject.toml
[project]
name = "dagster-and-etl"
version = "0.0.1"
description = "Dagster ETL"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dagster==1.11.8",
    "dagster-dg-cli",
    "dagster-webserver",
    "dagster-duckdb",
    "dagster-dlt",
    "dagster-sling",
    "psycopg2-binary",
    "dlt[duckdb]",
    "pandas",
    "azure-storage-blob>=12.26.0",
    "requests>=2.32.5",
    "pydantic>=2.11.7",
]

[tool.uv]
dev-dependencies = [
    "ruff",
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "dagster_and_etl"
registry_modules = [
    "dagster_and_etl.components.*",
]

[tool.ruff.lint]
ignore = [
    "ERA001", # comments are fine for a learning environment
    "E501",   # Line too long for docstrings
    "F401",   # unused imports already caught by pylance
]

fixable = ["ALL"]
unfixable = [
    "F401", # dont remove unused imports
]
